Update QueenFX.ino
wow, so after checking all of my electrical, and systematically getting back to the very root of my code I found out that the flickering issue came from a bug in the initializer for FastLED. I now set up each port (PORTD and PORTC) independently. Never would have guessed that. The crazy thing is that it must be a subtle timing thing...