Update QueenFX.ino
So much progress, but brain dead for the day. Renamed tubeLengths to tubeLength because I kept getting it wrong. Wrote a simple program to test colors, fixed up ShowLeds (mostly fixing off by one errors and simplifying a bit), DebugColors and fixed mirror and fillsolid to respect the tube lengths.